联动删除

+++ database/migrations/2022_05_14_115850_create_replies_table.php
@@ -4,14 +4,14 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateRepliesTable extends Migration 
+class CreateRepliesTable extends Migration
 {
 	public function up()
 	{
 		Schema::create('replies', function(Blueprint $table) {
             $table->increments('id');
             $table->integer('topic_id')->unsigned()->default(0)->index();
-            $table->bigInteger('user_id')->unsigned()->default(0)->index();
+            $table->integer('user_id')->unsigned()->default(0)->index();
             $table->text('content');
             $table->timestamps();
         });
